@@ -9,23 +9,24 @@
   },
   "dependencies": {
     "axios": "^0.21.1",
-    "bootstrap-vue": "^2.21.2",
+    "bootstrap-icons-vue": "^0.7.0",
     "core-js": "^3.12.1",
+    "mitt": "^2.1.0",
     "sass": "^1.32.8",
     "sass-loader": "^10.1.1",
     "sweetalert2": "^10.16.9",
-    "vue": "^2.6.11",
-    "vue-router": "^3.5.1",
-    "vuex": "^3.6.2"
+    "vue": "^3.0.0-beta.1",
+    "vue-router": "^4.0.8",
+    "vuex": "^4.0.0-alpha.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0-alpha.0"
   },
   "eslintConfig": {
     "root": true,
@@ -33,7 +34,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
