@@ -1,2 +1,5 @@
-import Vue from 'vue';
-export const EventBus = new Vue();
+//import Vue from 'vue';
+//export const EventBus = new Vue();
+
+import mitt from 'mitt';
+export const EventBus = mitt();
